Additions -Repo with Dao and Network Objects -View Model and Live Data Support for all necessary activities -New SQL queries -Conversion Class for Date Operation Fixes -Shared Pref with Application Context -Queries using String and Date Time -Repo Function and Network calls fixed Issues -Day Detail SQL Query returning Null -Icon not rendered -OverFlowing Text on Main Screen

+++ app/src/main/java/com/example/a3/testapp/DataModel/Maximum.java
@@ -6,18 +6,20 @@ import com.google.gson.annotations.SerializedName;
 
 public class Maximum{
 
+
+
 	@SerializedName("Value")
-	private int value;
+	private double value;
 
-	public Maximum(int value) {
+	public Maximum(double value) {
 		this.value = value;
 	}
 
-	public void setValue(int value){
+	public void setValue(double value){
 		this.value = value;
 	}
 
-	public int getValue(){
+	public double getValue(){
 		return value;
 	}
